+++ src/main/java/com/training/java/transaction/tracker/service/request/DeleteTransactionRequest.java
@@ -2,13 +2,13 @@ package com.training.java.transaction.tracker.service.request;
 
 public class DeleteTransactionRequest {
 
-  private final int transactionId;
+  private final Long transactionId;
 
-  public DeleteTransactionRequest(int transactionId) {
+  public DeleteTransactionRequest(Long transactionId) {
     this.transactionId = transactionId;
   }
 
-  public int getTransactionId() {
+  public Long getTransactionId() {
     return transactionId;
   }
 }
